+++ ast/ArgParsing.h
@@ -2,17 +2,12 @@
 #define SORBET_AST_ARG_PARSING_H
 #include "ast/ast.h"
 namespace sorbet::ast {
-struct ParsedArg {
-    core::LocOffsets loc;
-    core::LocalVariable local;
-    core::ArgInfo::ArgFlags flags;
-};
 class ArgParsing {
 public:
-    static std::vector<ParsedArg> parseArgs(const ast::MethodDef::ARGS_store &args);
-    static std::vector<uint32_t> hashArgs(core::Context ctx, const std::vector<ParsedArg> &args);
+    static std::vector<core::ParsedArg> parseArgs(const ast::MethodDef::ARGS_store &args);
+    static std::vector<uint32_t> hashArgs(core::Context ctx, const std::vector<core::ParsedArg> &args);
     // Returns the default argument value for the given argument, or nullptr if not specified. Mutates arg.
-    static ExpressionPtr getDefault(const ParsedArg &parsedArg, ExpressionPtr arg);
+    static ExpressionPtr getDefault(const core::ParsedArg &parsedArg, ExpressionPtr arg);
 };
 }; // namespace sorbet::ast
